Extract interfaces
