+++ package.json
@@ -5,7 +5,7 @@
   "scripts": {
     "dev": "vite --host",
     "build": "vue-tsc --noEmit && vite build",
-    "build2": "vite build",
+    "check": "vue-tsc --noEmit",
     "preview": "vite preview",
     "prepare": "husky install"
   },
