+++ src/enums/index.ts
@@ -1,0 +1,5 @@
+export enum MessageType {
+  Question = 0,
+  Answer = 1,
+  Note = 2,
+}
