+++ src/interfaces/index.ts
@@ -1,0 +1,10 @@
+export * from "./chat";
+export * from "./complaint";
+export * from "./feedback";
+export * from "./login";
+export * from "./page";
+export * from "./question";
+export * from "./recommend";
+export * from "./role";
+export * from "./share";
+export * from "./user";
