+++ src/interfaces/login.ts
@@ -1,0 +1,4 @@
+export interface AccessToken {
+  access_token: string;
+  token_type: string;
+}
