+++ src/interfaces/question.ts
@@ -1,0 +1,7 @@
+type Counter = { [key: string]: int };
+
+export type QuestionStats = {
+  word: Counter;
+  intent: Counter;
+  entity: Counter;
+};
