+++ src/interfaces/share.ts
@@ -1,0 +1,9 @@
+export interface SharedLink {
+  create_time: string;
+  expire_time: string;
+  max_uses: int;
+  use_times: int;
+  readonly: boolean;
+  valid: boolean;
+  id: string;
+}
