+++ src/types/globals.d.ts
