+++ src/views/admin/chat/MessagePage.vue
@@ -84,7 +84,8 @@
 </template>
 
 <script setup lang="ts">
-import { FeedbackStats, getFeedbackStats } from "@/api/feedback";
+import { getFeedbackStats } from "@/api/feedback";
+import type { FeedbackStats } from "@/interfaces";
 
 import FeedbackChart from "./components/FeedbackChart.vue";
 import FeedbackGauge from "./components/FeedbackGauge.vue";
