+++ src/views/admin/chat/components/ShareTable.vue
@@ -67,9 +67,9 @@
 import { QTable } from "quasar";
 import { date } from "quasar";
 
-import { ChatSession, deleteSession, getAllSessions } from "@/api/chat";
-import { SharedLink, getAllShares, updateShare } from "@/api/share";
-import { TablePagination } from "@/typing/quasar";
+import { getAllShares, updateShare } from "@/api/share";
+import type { SharedLink } from "@/interfaces";
+import { TablePagination } from "@/types/quasar";
 import { formatDate, formatNow } from "@/utils/date-utils";
 import Message from "@/utils/message";
 import { addSSP, makeRequester } from "@/utils/paginating";
