+++ src/views/admin/user/UserPage.vue
@@ -45,7 +45,8 @@
 </template>
 
 <script setup lang="ts">
-import { UserStats, getUserStats } from "@/api/user";
+import { getUserStats } from "@/api/user";
+import type { UserStats } from "@/interfaces";
 
 import UserTable from "./components/UserTable.vue";
