+++ src/views/admin/user/components/ComplaintChart.vue
@@ -26,7 +26,7 @@ import { use } from "echarts/core";
 import { CanvasRenderer } from "echarts/renderers";
 import VChart from "vue-echarts";
 
-import type { ComplaintStats } from "@/api/complaint";
+import type { ComplaintStats } from "@/interfaces";
 
 use([
   TitleComponent,
